+++ silentphone2/support/zrtpcpp/srtp/CryptoContext.cpp
@@ -394,8 +394,8 @@ bool CryptoContext::checkReplay(uint16_t newSeq)
         }
 
         delta = -delta;
-        int idx = delta / 64;
-        uint64_t bit = 1UL << (delta % 64);
+        int idx = (int)delta / 64;
+        uint64_t bit = (uint64_t)1UL << (delta % 64);
         if ((replay_window[idx] & bit) == bit) {
             return false;  /* Packet already received ! */
         }
@@ -442,8 +442,8 @@ void CryptoContext::update(uint16_t newSeq)
     }
     else {
         delta = -delta;
-        int idx = delta / 64;
-        uint64_t bit = 1UL << (delta % 64);
+        int idx = (int)delta / 64;
+        uint64_t bit = (uint64_t)1UL << (delta % 64);
         replay_window[idx] |= bit;
     }
